src/main/webapp/app/entities/ha/route/ha-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IHa, Ha } from '../ha.model';
import { HaService } from '../service/ha.service';

import { HaRoutingResolveService } from './ha-routing-resolve.service';

describe('Service Tests', () => {
  describe('Ha routing resolve service', () => {
    let mockRouter: Router;
    let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
    let routingResolveService: HaRoutingResolveService;
    let service: HaService;
    let resultHa: IHa | undefined;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        providers: [Router, ActivatedRouteSnapshot],
      });
      mockRouter = TestBed.inject(Router);
      mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
      routingResolveService = TestBed.inject(HaRoutingResolveService);
      service = TestBed.inject(HaService);
      resultHa = undefined;
    });

    describe('resolve', () => {
      it('should return IHa returned by find', () => {
        // GIVEN
        service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultHa = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultHa).toEqual({ id: 123 });
      });

      it('should return new IHa if id is not provided', () => {
        // GIVEN
        service.find = jest.fn();
        mockActivatedRouteSnapshot.params = {};

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultHa = result;
        });

        // THEN
        expect(service.find).not.toBeCalled();
        expect(resultHa).toEqual(new Ha());
      });

      it('should route to 404 page if data not found in server', () => {
        // GIVEN
        jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Ha })));
        mockActivatedRouteSnapshot.params = { id: 123 };

        // WHEN
        routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
          resultHa = result;
        });

        // THEN
        expect(service.find).toBeCalledWith(123);
        expect(resultHa).toEqual(undefined);
        expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
      });
    });
  });
});
